fix code for easy use and maintenance
- separate code for use shell command 2 core file
- make ui.py to display ui .exe file
