+++ cli/commands/common/i2c_command.py
@@ -1,0 +1,50 @@
+# command to get i2c information
+import ctypes
+
+def execute(device, command):
+    shell_command = f'i2cget -y 1 0x68 {command} w'
+    response = device.shell(shell_command)
+    return response
+
+def i2cStatus(device):
+    result = execute(device, '0x4a') 
+    i2c_status = result[0:6]
+    return i2c_status
+
+def input_voltage(device):
+    result = execute(device, '0x3b') 
+    input_v = result[0:6]
+    input_v = int(input_v, 16)*1.648
+    return input_v
+
+def input_current(device):
+    result = execute(device, '0x3e') 
+    input_c = result[0:6]
+    input_c = int(input_c, 16)*(0.00146487 /0.005)
+    return input_c
+
+def system_voltage(device):
+    result = execute(device, '0x3c') 
+    sys_v = result[0:6]
+    sys_v = int(sys_v, 16)*1.648
+    return sys_v
+
+def battery_voltage(device):
+    result = execute(device, '0x3a') 
+    bat_v = result[0:6]
+    bat_v = int(bat_v, 16)*0.192264
+    return bat_v
+
+def battery_cuurent(device):
+    result = execute(device, '0x3d') 
+    bat_c = result[0:6]
+    bat_c = int(bat_c, 16)
+    bat_c = ctypes.c_int16(bat_c).value
+    bat_c = bat_c*(0.00146487/0.003)
+    return bat_c
+
+def battery_percentage(device):
+    result = execute(device, '0x13') 
+    bat_p = result[0:6]
+    bat_p =((int(bat_p, 16)-16384)/32768)*100
+    return bat_p
